Rename module and request member, correct type
