+++ ruma-federation-api/src/query.rs
@@ -1,5 +1,5 @@
 //! Endpoints to retrieve information from a homeserver about a resource.
 
-pub mod custom;
+pub mod get_custom_information;
 pub mod get_profile_information;
 pub mod get_room_information;
