+++ ruma-federation-api/src/query/get_custom_information.rs
